Solucionado error con back

+++ BackEnd/AnalisisQuimicos.Core/Services/PDFGeneratorService.cs
@@ -204,7 +204,7 @@ namespace AnalisisQuimicos.Core.Services
             PaginaHTML_Texto = PaginaHTML_Texto.Replace("@CONT", contactos.Nombre);
             PaginaHTML_Texto = PaginaHTML_Texto.Replace("@ANA", analisis.Nombre);
             
-            Files file = _unidadDeTrabajo.FilesRepository.Download(usuario.Firma).Result;
+            Files file = _unidadDeTrabajo.FilesRepository.Download((int)usuario.Firma).Result;
 
             PaginaHTML_Texto = PaginaHTML_Texto.Replace("@Image", file.Path);
